requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
huggingface_hub==0.19.4
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
fastapi
uvicorn
python-multipart
